Shorten the overflow variable to shrink running time It runs overnigh to trigger the int overflow in FuzzBALL. To save time, change unsigned int to unsigned short

+++ c/bitvector-loops/overflow_1-2.c
@@ -8,7 +8,7 @@ void __VERIFIER_assert(int cond) {
 }
 
 int main(void) {
-  unsigned int x = 10;
+  unsigned short x = 10;
 
   while (x >= 10) {
     x += 2;
